app/backend/requirements.txt
aiodocker==0.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
check-jsonschema==0.33.3
click==8.3.0
click-default-group==1.2.4
click-didyoumean==0.3.1
cryptography==46.0.3
dateparser==1.2.2
distlib==0.4.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fasteners==0.19
filelock==3.19.1
firebase_admin==7.1.0
flatten-dict==0.4.2
frozenlist==1.7.0
git-filter-repo==2.47.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
meltano==3.9.1
msgpack==1.1.2
multidict==6.6.4
packaging==25.0
platformdirs==4.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
regress==2025.5.1
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
structlog==25.4.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.22
virtualenv==20.34.0
whitenoise==6.11.0
wrapt==1.17.3
yarl==1.20.1
